Added function tokens
