+++ tests/queries/0_stateless/2028_tokens.reference
@@ -1,0 +1,8 @@
+['test']
+['test1','test2','test3']
+['test1','test2','test3','test4']
+['test1','test2','test3','test4']
+['test']
+['test1','test2','test3']
+['test1','test2','test3','test4']
+['test1','test2','test3','test4']
