+++ tests/queries/0_stateless/2028_tokens.sql
@@ -1,0 +1,9 @@
+SELECT tokens('test');
+SELECT tokens('test1, test2, test3');
+SELECT tokens('test1, test2,     test3, test4');
+SELECT tokens('test1,;\ test2,;\ test3,;\   test4');
+
+SELECT tokens(materialize('test'));
+SELECT tokens(materialize('test1, test2, test3'));
+SELECT tokens(materialize('test1, test2,     test3, test4'));
+SELECT tokens(materialize('test1,;\ test2,;\ test3,;\   test4'));
